requirements.txt
anthropic
httpx
notion-client==2.3.0
pydantic
pymupdf
pyyaml
python-dotenv
pytest
